requirements.txt
pandas
numpy<2.0
Pillow
opencv-python
face_recognition
mysql-connector-python
bcrypt
tkcalendar